Triggers and Client Separation

+++ Source/WrightsAtHome/Server/API/Sensors/SensorReadingInfo.cs
@@ -5,7 +5,7 @@
         public int Id { get; set; }
         public string Name { get; set; }
         public string SensorType { get; set; }
-        public int Reading { get; set; }
+        public decimal Reading { get; set; }
         public string SmallImageUrl { get; set; }
         public string LargeImageUrl { get; set; }
     }
